Making Object3D classes consistent
issue: MatterHackers/MCCentral#3702
Consider adding 'Object3D' suffix to Object3D types or use consistent format
