+++ DesignTools/Operations/TranslateObject3D.cs
@@ -28,17 +28,25 @@ either expressed or implied, of the FreeBSD Project.
 */
 
 using MatterHackers.DataConverters3D;
-using MatterHackers.Localizations;
+using MatterHackers.VectorMath;
 
 namespace MatterHackers.MatterControl.DesignTools.Operations
 {
-	public class Group3D : Object3D
+	public class TranslateObject3D : Object3D
 	{
-		public Group3D()
+		public TranslateObject3D()
 		{
-			Name = "Group".Localize();
 		}
 
-		public override bool CanRemove => true;
+		public TranslateObject3D(IObject3D item, double x = 0, double y = 0, double z = 0)
+			: this(item, new Vector3(x, y, z))
+		{
+		}
+
+		public TranslateObject3D(IObject3D item, Vector3 translation)
+		{
+			Matrix *= Matrix4X4.CreateTranslation(translation);
+			Children.Add(item.Clone());
+		}
 	}
 }
