+++ MatterControl.csproj
@@ -92,16 +92,27 @@
     <Compile Include="DesignTools\Attributes\ShowUpdateButtonAttribute.cs" />
     <Compile Include="DesignTools\EditableTypes\DirectionAxis.cs" />
     <Compile Include="DesignTools\EditableTypes\DirectionVector.cs" />
-    <Compile Include="DesignTools\Operations\Group3D.cs" />
+    <Compile Include="DesignTools\Operations\Align2D.cs" />
+    <Compile Include="DesignTools\Operations\AlignObject3D.cs" />
+    <Compile Include="DesignTools\Operations\ArrayAdvancedObject3D.cs" />
+    <Compile Include="DesignTools\Operations\ArrayLinearObject3D.cs" />
+    <Compile Include="DesignTools\Operations\ArrayRadialObject3D.cs" />
+    <Compile Include="DesignTools\Operations\FitToBoundsObject3D.cs" />
+    <Compile Include="DesignTools\Operations\GroupObject3D.cs" />
     <Compile Include="DesignTools\Operations\Image\AlphaThresholdFunction.cs" />
     <Compile Include="DesignTools\Operations\Image\HueThresholdFunction.cs" />
-    <Compile Include="DesignTools\Operations\Image\ImageToPath.cs" />
+    <Compile Include="DesignTools\Operations\Image\ImageToPathObject3D.cs" />
     <Compile Include="DesignTools\Operations\Image\IThresholdFunction.cs" />
-    <Compile Include="DesignTools\Operations\Image\LinearExtrude.cs" />
+    <Compile Include="DesignTools\Operations\Image\LinearExtrudeObject3D.cs" />
     <Compile Include="DesignTools\Operations\Image\MapOnMaxIntensity.cs" />
     <Compile Include="DesignTools\Operations\Image\SilhouetteThresholdFunction.cs" />
-    <Compile Include="DesignTools\Operations\OperationSource.cs" />
+    <Compile Include="DesignTools\Operations\OperationSourceObject3D.cs" />
+    <Compile Include="DesignTools\Operations\Path\SmoothPathObject3D.cs" />
     <Compile Include="DesignTools\Operations\PinchObject3D.cs" />
+    <Compile Include="DesignTools\Operations\RotateObject3D.cs" />
+    <Compile Include="DesignTools\Operations\ScaleObject3D.cs" />
+    <Compile Include="DesignTools\Operations\SetCenterObject3D.cs" />
+    <Compile Include="DesignTools\Operations\TranslateObject3D.cs" />
     <Compile Include="DesignTools\Primitives\ComponentObject3D.cs" />
     <Compile Include="DesignTools\Primitives\CubeObject3D.cs" />
     <Compile Include="DesignTools\Primitives\HalfCylinderObject3D.cs" />
@@ -110,6 +121,7 @@
     <Compile Include="Library\Providers\MatterControl\LibraryCollectionContainer.cs" />
     <Compile Include="Library\Providers\MatterControl\PartHistoryContainer.cs" />
     <Compile Include="Library\Providers\MatterControl\RootHistoryContainer.cs" />
+    <Compile Include="Library\Widgets\InsertionGroupObject3D.cs" />
     <Compile Include="PartPreviewWindow\GCodeDetails\BoolOption.cs" />
     <Compile Include="PartPreviewWindow\GCodeDetails\GCodeLayerDetailsView.cs" />
     <Compile Include="PartPreviewWindow\GCodeDetails\GCodeOptionsPanel.cs" />
@@ -122,6 +134,7 @@
     <Compile Include="PartPreviewWindow\StartPage\StartTabPage.cs" />
     <Compile Include="PartPreviewWindow\View3D\Actions\CombineObject3D.cs" />
     <Compile Include="PartPreviewWindow\View3D\Actions\IntersectionObject3D.cs" />
+    <Compile Include="PartPreviewWindow\View3D\Actions\ModifiedMeshObject3D.cs" />
     <Compile Include="PartPreviewWindow\View3D\Actions\SubtractAndReplaceObject3D.cs" />
     <Compile Include="PartPreviewWindow\View3D\Actions\SubtractObject3D.cs" />
     <Compile Include="PrinterControls\ControlWidgets\CalibrationControls.cs" />
@@ -150,16 +163,8 @@
     <Compile Include="DesignTools\Interfaces\IEditorDraw.cs" />
     <Compile Include="DesignTools\Interfaces\IPropertyGridModifier.cs" />
     <Compile Include="DesignTools\Lithophane.cs" />
-    <Compile Include="DesignTools\Operations\Align3D.cs" />
-    <Compile Include="DesignTools\Operations\ArrayAdvanced3D.cs" />
-    <Compile Include="DesignTools\Operations\ArrayLinear3D.cs" />
-    <Compile Include="DesignTools\Operations\ArrayRadial3D.cs" />
     <Compile Include="DesignTools\Operations\CurveObject3D.cs" />
-    <Compile Include="DesignTools\Operations\FitToBounds3D.cs" />
     <Compile Include="DesignTools\Operations\Object3DExtensions.cs" />
-    <Compile Include="DesignTools\Operations\Rotate.cs" />
-    <Compile Include="DesignTools\Operations\Scale.cs" />
-    <Compile Include="DesignTools\Operations\Path\SmoothPath.cs" />
     <Compile Include="DesignTools\Primitives\BaseObject3D.cs" />
     <Compile Include="DesignTools\Primitives\ImageObject3D.cs" />
     <Compile Include="DesignTools\Primitives\RingObject3D.cs" />
@@ -172,8 +177,6 @@
     <Compile Include="DesignTools\Primitives\TorusObject3D.cs" />
     <Compile Include="DesignTools\PublicPropertyEditor.cs" />
     <Compile Include="DesignTools\Primitives\ConeObject3D.cs" />
-    <Compile Include="DesignTools\Operations\SetCenter.cs" />
-    <Compile Include="DesignTools\Operations\Translate.cs" />
     <Compile Include="DesignTools\Attributes\SortableAttribute.cs" />
     <Compile Include="DesignTools\Attributes\UnlockLinkAttribute.cs" />
     <Compile Include="Library\Providers\FileSystem\McxContainer.cs" />
@@ -224,7 +227,6 @@
     <Compile Include="Library\Providers\MatterControl\PrintHistoryItem.cs" />
     <Compile Include="Library\Providers\WritableContainer.cs" />
     <Compile Include="Library\Widgets\ExpandCheckboxButton.cs" />
-    <Compile Include="Library\Widgets\InsertionGroup.cs" />
     <Compile Include="PartPreviewWindow\NewTabButton.cs" />
     <Compile Include="PartPreviewWindow\PartTabPage.cs" />
     <Compile Include="PartPreviewWindow\StartPage\ExploreItem.cs" />
@@ -238,7 +240,6 @@
     <Compile Include="PartPreviewWindow\SectionWidget.cs" />
     <Compile Include="PartPreviewWindow\SliceLayerSelector.cs" />
     <Compile Include="PartPreviewWindow\Toolbar.cs" />
-    <Compile Include="PartPreviewWindow\View3D\Actions\MeshWrapper.cs" />
     <Compile Include="PartPreviewWindow\View3D\PrinterBar\CancelButton.cs" />
     <Compile Include="PartPreviewWindow\View3D\PrinterBar\PauseResumeButton.cs" />
     <Compile Include="Library\ReadOnlyStreamItem.cs" />
